requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
beautifulsoup4==4.9.2
requests==2.24.0
lxml==4.5.2

##
Werkzeug==0.16.1

#Test
unittest-xml-reporting==2.5.1
coverage==4.5.4
nose==1.3.7
nose-htmloutput==0.6.0
